+++ src/main/java/com/example/service/PublisherServiceImpl.java
@@ -25,7 +25,7 @@ public class PublisherServiceImpl implements PublisherService{
     }
 
     @Override
-    public Publisher findById(int id) {
+    public Publisher findById(Long id) {
         Optional<Publisher> publishers = publisherRepository.findById(id);
         return publishers.orElseGet(() -> new Publisher("Go Away", 000));
     }
